requirements.txt
altgraph==0.16.1
Click==7.0
colorama==0.4.1
future==0.17.1
macholib==1.11
pefile==2018.8.8
Pillow==5.3.0
PyInstaller==3.4
